% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birthwt}
\alias{birthwt}
\title{Risk Factors Associated with Low Infant Birth Weight}
\format{A data frame with 189 rows and 19 variables.  First 10 columns documented in \code{\link[MASS]{birthwt}}.  Additional 9 columns defined in Claeskens and Hjort (2008).
\describe{
  \item{}{}
  \item{}{}
  ...
}}
\source{
\pkg{MASS} package (Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer); originally from Hosmer, D.W. and Lemeshow, S. (1989) Applied Logistic Regression. New York: Wiley
}
\usage{
birthwt
}
\description{
Risk Factors Associated with Low Infant Birth Weight
}
\references{
Claeskens, G., & Hjort, N. L. (2008). Model selection and model averaging (Vol. 330). Cambridge: Cambridge University Press.
}
\keyword{datasets}
