% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fic.coxph.R
\name{fic.coxph}
\alias{fic.coxph}
\title{Focused Information Criterion for Cox proportional hazard regression models}
\usage{
\method{fic}{coxph}(wide, inds, inds0 = NULL, gamma0 = 0, focus, X = NULL,
  t = NULL, sub = NULL, tidy = TRUE)
}
\arguments{
\item{wide}{Fitted model object containing the wide model.}

\item{inds}{matrix or vector of indicators for which parameters are included in the submodel or submodels to be assessed.

A matrix should be supplied if multiple submodels are to be assessed.  This should have number of rows equal to the number of submodels to be assessed, and number of columns equal to the total number of parameters in the wide model.  It contains 1s in the positions where the parameter is included in the submodel, and 0s in positions where the parameter is excluded.  This should always be 1 in the positions defining the narrow model, as specified in `inds0`.}

\item{inds0}{Vector of indicators specifying the narrow model, in the same format as \code{inds}.  If this is omitted, the narrow model is assumed to be defined by the first row of \code{inds} (if \code{inds} is a matrix), or \code{inds} itself if this is a vector.}

\item{gamma0}{Vector of special values taken by the parameters \eqn{gamma} which define the narrow model.

This defaults to 0, as in covariate selection, where excluded coefficients are fixed to 0. 

This should either be a scalar, assumed to be the same for all parameters fixed in the narrow model, or a vector of length equal to the number of parameters from the wide model which are fixed in the narrow model, that is, the number of entries of inds0 which are zero.}

\item{focus}{\code{"hr"} for the hazard ratio between individuals with covariate values of \code{X} and 0. 

\code{"survival"} for the survival probability at time or times given in \code{t}.

\code{"cumhaz"} for the cumulative hazart at time or times given in \code{t}.}

\item{X}{Second argument of the focus function, typically giving covariate values defining the focus. This can either be a matrix or a vector.

This is optional.  If not supplied, \code{focus} should return a scalar giving the focus value at parameters \code{par}.

To compute focused model comparison statistics for the same focus function evaluated at multiple covariate values, \code{X} should be a matrix, with number of columns equal to the number of parameters in the wide model, and number of rows equal to the number of alternative covariate values.

For a typical regression model, the first parameter will denote an intercept, so the first value of \code{X} should be 1, and the remaining values should correspond to covariates whose coefficients form parameters of the wide model.  See the examples in the vignette.

If just one covariate value is needed, then \code{X} can be a vector of length equal to the number of parameters in the wide model.}

\item{t}{times to evaluate the focus at.  Only relevant for survival and cumulative hazard focuses, as the hazard ratio is constant through time.}

\item{sub}{List of fitted model objects for each submodel to be assessed.  This is optional, and only required if you want the estimate of the focus function under each submodel to be included in the results.}

\item{tidy}{If \code{TRUE} return the results as a data frame.  If \code{FALSE}, return the results as a three-dimensional array, with dimensions indexed by the submodels, result statistics and \code{X} values respectively.}
}
\description{
Focused information criterion for Cox models fitted with \code{coxph} from the \code{survival} package.  Only three built-in focuses are handled, with currently no support for user-defined focuses.  This has not been fully tested and is not yet recommended for use.
}
