% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fic.flexsurvreg.R
\name{fic.flexsurvreg}
\alias{fic.flexsurvreg}
\title{Focused Information Criterion for flexible parametric survival models}
\usage{
\method{fic}{flexsurvreg}(wide, sub = NULL, inds, inds0, gamma0 = 0,
  focus = NULL, focus_deriv = NULL, X = NULL, ...)
}
\arguments{
\item{wide}{Object of class \code{\link{"flexsurvreg"}} containing the wide model.
These are returned, for example, by the \code{\link{flexsurvreg}} and 
the \code{\link{flexsurvspline}} functions.}

\item{sub}{Object of class \code{\link{flexsurvreg}} containing the submodel to be assessed.}

\item{inds}{TODO better documentation for what indices correspond to what parameters, and what model selection problems are supported.
Do at least covariate selection, selection between gamma, generalised gamma etc.
Clarify choice between splines with different numbers of knots not supported, unless defined specially with nested knot locations.}

\item{inds0}{Specification of narrow model, in the same format as \code{inds}.  TODO error checking}

\item{gamma0}{Vector of special values taken by the parameters \eqn{gamma} which define the narrow model.
Defaults to all 0, as in covariate selection, where excluded coefficients are fixed to 0. 
Either a scalar, assumed to be the same for all parameters fixed in the narrow model,
or a vector of length equal to the number of parameters from the wide model which are fixed in the narrow model.}

\item{focus}{An R function whose first argument is a vector of parameters; and which returns one scalar: the focus quantity of interest.  Not required if \code{focus_deriv} is specified.

Alternatively, this can be a character string naming a built-in focus function supplied by the \pkg{fic} package.  See \code{\link{focus_fns}}.}

\item{focus_deriv}{Vector of partial derivatives of the focus function with respect to the parameters in the wide model.  If not supplied, this is computed from the function supplied in \code{focus}, using numerical differentiation.}

\item{X}{second argument of the focus function.  This will typically be a vector of covariate values.}

\item{...}{Other arguments to the focus function can be supplied here.  Currently no examples of this.}
}
\description{
Focused information criterion for parametric survival models fitted with the \pkg{flexsurv} package.
}
