% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fic.glm.R
\name{fic.glm}
\alias{fic.glm}
\title{Focused Information Criterion for generalized linear models}
\usage{
fic.glm(object, inds, pp, focus = NULL, focus.deriv = NULL)
}
\arguments{
\item{object}{Object returned by \code{\link{glm}} containing the fitted model}

\item{inds}{Vector of 0s and 1s of length \code{length(ests) - pp}, with 1s in the positions which define the submodel to be compared against the wide model.}

\item{pp}{Number of parameters which we would always include in any submodel.  These parameters are assumed to be the first \code{pp} parameters included in \code{ests}.}

\item{focus}{An R function taking one argument: a vector of parameters; and returning one scalar: the focus quantity of interest.  Not required if \code{focus.deriv} is specified}

\item{focus.deriv}{Vector of partial derivatives of the focus function with respect to the parameters in the wide model.  If not supplied, this is computed from the function supplied in \code{focus}, using numerical differentiation.}
}
\description{
Focused information criterion for generalized linear models fitted with glm()
}
