% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fic.R
\name{fic}
\alias{fic}
\alias{fic.default}
\alias{FIC}
\title{Focused Information Criterion: generic function}
\usage{
fic(x, ...)

\method{fic}{default}(par, J, inds, inds0, gamma0 = 0, n, focus = NULL,
  focus_deriv = NULL, X = NULL, parsub = NULL, ...)
}
\arguments{
\item{\dots}{Other arguments to the focus function can be supplied here.  Currently no examples of this.}

\item{par}{Vector of maximum likelihood estimates from the wide model}

\item{J}{Information matrix from the wide model, evaluated at the maximum likelihood estimates and divided by \code{n}.}

\item{inds}{Vector of 0s and 1s of same length as \code{par}, with 1s in the positions where the parameters of the wide model are included in the submodel to be assessed, and 0s elsewhere.}

\item{inds0}{Specification of narrow model, in the same format as \code{inds}.  TODO error checking}

\item{gamma0}{Vector of special values taken by the parameters \eqn{gamma} which define the narrow model.
Defaults to all 0, as in covariate selection, where excluded coefficients are fixed to 0. 
Either a scalar, assumed to be the same for all parameters fixed in the narrow model,
or a vector of length equal to the number of parameters from the wide model which are fixed in the narrow model.}

\item{n}{Number of observations in the data used to fit the wide model.  TODO is this really needed for the MSE, or does it cancel?  Only of academic interest to compute FIC/error for the sqrt n transformed quantity?}

\item{focus}{An R function whose first argument is a vector of parameters; and which returns one scalar: the focus quantity of interest.  Not required if \code{focus_deriv} is specified.

Alternatively, this can be a character string naming a built-in focus function supplied by the \pkg{fic} package.  See \code{\link{focus_fns}}.}

\item{focus_deriv}{Vector of partial derivatives of the focus function with respect to the parameters in the wide model.  If not supplied, this is computed from the function supplied in \code{focus}, using numerical differentiation.}

\item{X}{second argument of the focus function.  This will typically be a vector of covariate values.}

\item{parsub}{Vector of maximum likelihood estimates from the submodel.
Only required to return the estimate of the focus quantity alongside 
the model assessment statistics for the submodel. If omitted, the estimate is omitted.

This can be a vector with \code{nc} elements, e.g. one for each covariate coefficient in the wide model.

Alternatively this can be a matrix with \code{nf} rows and \code{nc} columns. 
The FIC calculation is then "vectorised".  With one call of the \code{fic} function, the   
FIC and related statistics are calculated \code{nf} times, with 
the focus function evaluated at each of \code{nf} alternative 
covariate values. 

Not yet tested in situations other than supplying covariate values.}
}
\value{
A vector containing the following components, describing characteristics of the defined submodel (references in Chapter 6 of Claeskens and Hjort, 2008)

\item{FIC}{The focused information criterion (equation 6.1). }

\item{rmse}{The root mean square error of the estimate of the focus quantity.  Defined on page 150 (equation 6.7).}

\item{rmse.adj}{The root mean square error, with an adjustment to avoid negative squared bias.  Defined on page 157 as the sum of the variance and the squared adjusted bias.}

\item{bias}{The estimated bias of the focus quantity (unadjusted).  Defined on page 157.}

\item{bias.adj}{The estimated bias of the focus quantity (adjusted to avoid negative squared bias).  This is defined as the square root of the quantity "sqb3(S)", page 152, multiplied by the sign of the unadjusted bias, and divided by the square root of the sample size. }

\item{se}{The estimated standard error (root variance) of the focus quantity.  Defined on page 157.}
}
\description{
Focused Information Criterion: generic function

Focused information criterion for general models.  These methods estimate the bias and variance of estimates of a quantity of interest (the "focus") when parameters are excluded from an "wide" model that is assumed to generate the data.
}
\examples{
TODO

}
\references{
Claeskens, G., & Hjort, N. L. (2008). Model selection and model averaging (Vol. 330). Cambridge: Cambridge University Press.

Claeskens, G., & Hjort, N. L. (2003). The focused information criterion. Journal of the American Statistical Association, 98(464), 900-916.
}
\keyword{models}
