% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fic.R
\name{fic_core}
\alias{fic_core}
\alias{fic_multi}
\title{Focused Information Criterion: core calculation functions}
\usage{
fic_core(par, J, inds, inds0, gamma0 = 0, n, focus = NULL,
  focus_deriv = NULL, ...)

fic_multi(par, J, inds, inds0, gamma0 = 0, n, focus = NULL,
  focus_deriv = NULL, X = NULL, parsub = NULL, ...)
}
\arguments{
\item{par}{Vector of maximum likelihood estimates from the wide model}

\item{J}{Information matrix from the wide model, evaluated at the maximum likelihood estimates and divided by \code{n}.}

\item{inds}{Vector of 0s and 1s of same length as \code{par}, with 1s in the positions where the parameters of the wide model are included in the submodel to be assessed, and 0s elsewhere.   TODO DOC AND TEST FOR MATRIX}

\item{inds0}{Vector of indicators specifying the narrow model, in the same format as \code{inds}.  If this is omitted, the narrow model is assumed to be defined by the first row of \code{inds} (if \code{inds} is a matrix), or \code{inds} itself if this is a vector.}

\item{gamma0}{Vector of special values taken by the parameters \eqn{gamma} which define the narrow model.

This defaults to 0, as in covariate selection, where excluded coefficients are fixed to 0. 

This should either be a scalar, assumed to be the same for all parameters fixed in the narrow model, or a vector of length equal to the number of parameters from the wide model which are fixed in the narrow model, that is, the number of entries of inds0 which are zero.}

\item{n}{Number of observations in the data used to fit the wide model.  TODO is this really needed for the MSE, or does it cancel?  Only of academic interest to compute FIC/error for the sqrt n transformed quantity?}

\item{focus}{An R function with:

\itemize{
\item first argument named \code{par}, denoting a vector of parameters

\item an optional second argument named \code{X}, typically denoting covariate values.  The required format is documented below. 
}

The function should return the focus quantity of interest.  If \code{X} is supplied and has multiple rows, then \code{focus} should return a vector giving the focus for each row of \code{X}.  Otherwise \code{focus} should return a scalar giving the focus value at \code{par}.

Not required if \code{focus_deriv} is specified.

Alternatively, \code{focus} can be a character string naming a built-in focus function supplied by the \pkg{fic} package.  See \code{\link{focus_fns}}.}

\item{focus_deriv}{Vector of partial derivatives of the focus function with respect to the parameters in the wide model.  This is required by \code{fic.core}.

If there are multiple submodels, this should be a matrix with number of rows equal to the number of submodels, and number of columns equal to the number of parameters in the wide model.  If there is a single submodel, this should be a vector with number of columns equal to the number of parameters in the wide model.

This should take the value given by \code{gamma0} in positions where the parameter is excluded from the submodel.
For example, coefficients of covariates should take the value 0 if the covariate is excluded.}

\item{...}{Other arguments to the focus function can be supplied here.  Currently no examples of this.}

\item{X}{Second argument of the focus function, typically giving covariate values defining the focus. This can either be a matrix or a vector.

This is optional.  If not supplied, \code{focus} should return a scalar giving the focus value at parameters \code{par}.

To compute focused model comparison statistics for the same focus function evaluated at multiple covariate values, \code{X} should be a matrix, with number of columns equal to the number of parameters in the wide model, and number of rows equal to the number of alternative covariate values.

For a typical regression model, the first parameter will denote an intercept, so the first value of \code{X} should be 1, and the remaining values should correspond to covariates whose coefficients form parameters of the wide model.  See the examples in the vignette.

If just one covariate value is needed, then \code{X} can be a vector of length equal to the number of parameters in the wide model.}

\item{parsub}{Vector of maximum likelihood estimates from the submodel.
Only required to return the estimate of the focus quantity alongside 
the model assessment statistics for the submodel. If omitted, the estimate is omitted.}
}
\value{
See \code{\link{fic}}
}
\description{
Core FIC calculation functions underlying the user interface in \code{\link{fic}}.
\code{fic_core} just handles one submodel, while \code{fic_multi} can assess multiple submodels of the same wide model.  For \code{fic_multi}, \code{\link{inds}} and \code{\link{parsub}} can be matrices with one row per submodel, while for \code{fic_core} they must be vectors.
}
