% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{fit_submodels}
\alias{fit_submodels}
\title{Fit submodels of a wide model, defined by a matrix of indicators
for inclusion of covariates.}
\usage{
fit_submodels(wide, inds)
}
\arguments{
\item{wide}{Fitted model object containing the wide model.}

\item{inds}{Matrix or vector of indicators for which parameters are included in the submodel or submodels to be assessed.

A matrix should be supplied if there are multiple submodels.  This should have number of rows equal to the number of submodels, and number of columns equal to the total number of parameters in the wide model.  It contains 1s in the positions where the parameter is included in the submodel, and 0s in positions where the parameter is excluded.  This should always be 1 in the positions defining the narrow model, as specified in \code{inds0}.}
}
\value{
List of all fitted submodel objects
}
\description{
Fit the submodels of a wide model \code{wide} which are defined by
\code{inds}.  This can only be used for covariate selection
problems, where the submodels contain different subsets of
covariates.
}
\details{
Requires \code{wide} to have a component named
\code{call} giving the function call used to produce \code{wide}.
This call should include a \code{formula} component, which this
function updates in order to define and fit the submodel.  This
should work for most standard linear-type models in common R packages.
}
