% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fic.coxph.R
\name{get_H0}
\alias{get_H0}
\title{Return baseline cumulative hazard from fitted model at given times
Assumes hazard is piecewise constant, thus cumulative hazard is piecewise linear}
\usage{
get_H0(H0, t)
}
\description{
Return baseline cumulative hazard from fitted model at given times
Assumes hazard is piecewise constant, thus cumulative hazard is piecewise linear
}
