% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fic.R
\name{ggplot_fic}
\alias{ggplot_fic}
\title{Plot focused model comparison statistics: ggplot2 method}
\usage{
ggplot_fic(x, ylab = NULL, xlab = NULL)
}
\arguments{
\item{x}{Output from \code{\link{fic}}.}

\item{ylab}{y=axis label.}

\item{xlab}{x-axis label.}
}
\description{
Plot focused model comparison statistics: ggplot2 method
}
