% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{melanoma}
\alias{melanoma}
\title{Melanoma}
\format{TODO
\describe{
  \item{}{}
  \item{}{}
  ...
}}
\source{
TODO
}
\usage{
melanoma
}
\description{
Melanoma
}
\references{
Claeskens, G., & Hjort, N. L. (2008). Model selection and model averaging (Vol. 330). Cambridge: Cambridge University Press.
}
\keyword{datasets}
