% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{newdata_to_X}
\alias{newdata_to_X}
\title{Convert `newdata` to a design matrix excluding the intercept}
\usage{
newdata_to_X(newdata, wide)
}
\description{
Numerics are allowed for character factor levels
}
