% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fic.R
\name{plot.fic}
\alias{plot.fic}
\title{What is plotted depends on whether `fic` was called with a `sub`
argument so that the focus estimates are available in the object
returned by `fic`.}
\usage{
\method{plot}{fic}(x, ylab = NULL, xlab = NULL, pch = 19, mfrow = NULL,
  ...)
}
\description{
If the focus estimates are available, then the focus estimates are
plotted against the root MSE.  One plot is made for each covariate
value defining different focuses.
}
\details{
If the focus estimates are unavailable, then the standard errors
of the focus estimate are plotted against the corresponding bias.
The plot points are shaded with darkness proportional to the RMSE,
with the point of maximum RMSE in black.
}
