% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fic.R
\name{expand_inds}
\alias{expand_inds}
\title{Convert inds, inds0 from model.frame (newdata) dimensions to model.matrix (X) dimensions
given a wide model
Replicates 0s or 1s for each factor term to number of parameters for that factor
Needs model.matrix() function to work on the wide model
or else they'll have to use lower level FIC function for the moment
or not have factors in their model. 
Converts vector, data frame or matrix input to matrix output}
\usage{
expand_inds(inds, wide, name = "inds")
}
\description{
Convert inds, inds0 from model.frame (newdata) dimensions to model.matrix (X) dimensions
given a wide model
Replicates 0s or 1s for each factor term to number of parameters for that factor
Needs model.matrix() function to work on the wide model
or else they'll have to use lower level FIC function for the moment
or not have factors in their model. 
Converts vector, data frame or matrix input to matrix output
}
