% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fic-package.R
\docType{package}
\name{fic-package}
\alias{fic-package}
\alias{_PACKAGE}
\title{fic: Focused model comparison}
\description{
TODO
}
\section{ref vignettes}{
NA
}

\references{
TODO
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chjackson/fic}
  \item Report bugs at \url{https://github.com/chjackson/fic/issues}
}

}
\author{
Christopher Jackson \email{chris.jackson@mrc-bsu.cam.ac.uk}
}
\keyword{package}
