% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fic-package.R
\docType{package}
\name{fic-package}
\alias{fic-package}
\title{Focused Information Criteria for Model Comparison}
\description{
For a full explanation of the methods behind the \pkg{fic} package and worked examples of using it, see the main package vignette "Different models for different purposes: focused model comparison in R"
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chjackson/fic}
  \item Report bugs at \url{https://github.com/chjackson/fic/issues}
}

}
\author{
\strong{Maintainer}: Christopher Jackson \email{chris.jackson@mrc-bsu.cam.ac.uk} (package design and programming)

Authors:
\itemize{
  \item Gerda Claeskens \email{gerda.claeskens@kuleuven.be} (method development and design advice)
}

Other contributors:
\itemize{
  \item Howard Thom \email{howard.thom@bristol.ac.uk} [contributor]
}

}
