% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fic-package.R
\docType{package}
\name{fic-package}
\alias{fic-package}
\title{Focused Information Criteria for Model Comparison}
\description{
For a full explanation of the methods behind the \pkg{fic} package and worked examples of using it, see the main package vignette "Different models for different purposes: focused model comparison in R"
}
