% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fic.glm.R
\name{fic.glm}
\alias{fic.glm}
\title{Focused Information Criterion for generalized linear models}
\usage{
\method{fic}{glm}(wide, sub = NULL, inds, inds0, focus = NULL,
  focus_deriv = NULL, X = NULL, ...)
}
\arguments{
\item{wide}{Object returned by \code{\link{glm}} containing the wide model.}

\item{sub}{Object returned by \code{\link{glm}} containing the submodel to be assessed.  Optional. 
Only required if you want the estimate of the focus function under the submodel to be included in the results.}

\item{inds}{Vector of 0s and 1s of same length as \code{par}, with 1s in the positions where the parameters of the wide model are included in the submodel to be assessed, and 0s elsewhere.}

\item{inds0}{Specification of narrow model, in the same format as \code{inds}.  TODO error checking}

\item{focus}{An R function whose first argument is a vector of parameters; and which returns one scalar: the focus quantity of interest.  Not required if \code{focus_deriv} is specified.

Alternatively, this can be a character string naming a built-in focus function supplied by the \pkg{fic} package.  See \code{\link{focus_fns}}.}

\item{focus_deriv}{Vector of partial derivatives of the focus function with respect to the parameters in the wide model.  If not supplied, this is computed from the function supplied in \code{focus}, using numerical differentiation.}

\item{X}{second argument of the focus function.  This will typically be a vector of covariate values.}

\item{...}{Other arguments to the focus function can be supplied here.  Currently no examples of this.}
}
\description{
Focused information criterion for generalized linear models fitted with glm().
This can be used to compare models with different covariates.  TODO any other model selection problems?
}
