% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fic.msm.R
\name{fic.msm}
\alias{fic.msm}
\title{Focused Information Criterion for multi-state models for panel data}
\usage{
\method{fic}{msm}(wide, sub = NULL, inds, inds0, focus = NULL,
  focus_deriv = NULL, X = NULL, ...)
}
\arguments{
\item{wide}{Object returned by \code{\link{msm}} containing the wide model.}

\item{sub}{Object returned by \code{\link{msm}} containing the submodel to be assessed.}

\item{inds}{TODO better documentation for what indices correspond to what parameters, and what model selection problems are supported.  Just covariate selection?  what about constraints?}

\item{inds0}{Specification of narrow model, in the same format as \code{inds}.  TODO error checking}

\item{focus}{An R function whose first argument is a vector of parameters; and which returns one scalar: the focus quantity of interest.  Not required if \code{focus_deriv} is specified.

Alternatively, this can be a character string naming a built-in focus function supplied by the \pkg{fic} package.  See \code{\link{focus_fns}}.}

\item{focus_deriv}{Vector of partial derivatives of the focus function with respect to the parameters in the wide model.  If not supplied, this is computed from the function supplied in \code{focus}, using numerical differentiation.}

\item{X}{second argument of the focus function.  This will typically be a vector of covariate values.}

\item{...}{Other arguments to the focus function can be supplied here.  Currently no examples of this.}
}
\description{
Focused information criterion for multi-state models fitted with msm().  This is not yet a S3 method, but will probably be made into one
}
