% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fic.R
\name{fic}
\alias{fic}
\alias{FIC}
\title{Focused Information Criterion: general function}
\usage{
fic(ests, J, inds, pp, n, focus = NULL, focus.deriv = NULL)
}
\arguments{
\item{ests}{Vector of maximum likelihood estimates from the wide model}

\item{J}{Information matrix from the wide model, evaluated at the maximum likelihood estimates and divided by \code{n}.}

\item{inds}{Vector of 0s and 1s of length \code{length(ests) - pp}, with 1s in the positions which define the submodel to be compared against the wide model.}

\item{pp}{Number of parameters which we would always include in any submodel.  These parameters are assumed to be the first \code{pp} parameters included in \code{ests}.}

\item{n}{Number of observations in the data used to fit the wide model.}

\item{focus}{An R function taking one argument: a vector of parameters; and returning one scalar: the focus quantity of interest.  Not required if \code{focus.deriv} is specified}

\item{focus.deriv}{Vector of partial derivatives of the focus function with respect to the parameters in the wide model.  If not supplied, this is computed from the function supplied in \code{focus}, using numerical differentiation.}
}
\value{
A vector containing the focused information criterion, bias and variance quantities (TO DEFINE) for the defined submodel.
}
\description{
Focused information criterion for general models.
}
\examples{
TODO

}
\references{
Claeskens, G., & Hjort, N. L. (2008). Model selection and model averaging (Vol. 330). Cambridge: Cambridge University Press.

Claeskens, G., & Hjort, N. L. (2003). The focused information criterion. Journal of the American Statistical Association, 98(464), 900-916.
}
\keyword{models}
