% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fic.R
\name{fic.default}
\alias{fic.default}
\alias{FIC}
\alias{fic}
\title{Focused Information Criterion: main user interface}
\usage{
\method{fic}{default}(wide, inds, inds0, gamma0 = 0, focus = NULL,
  focus_deriv = NULL, X = NULL, sub = NULL, fns = NULL, tidy = TRUE,
  ...)

fic(x, ...)
}
\arguments{
\item{wide}{Fitted model object containing the wide model.}

\item{inds}{matrix or vector of indicators for which parameters are included in the submodel or submodels.

A matrix should be supplied if multiple submodels are to be assessed.  This should have number of rows equal to the number of submodels to be assessed, and number of columns equal to the total number of parameters in the wide model.  It contains 1s in the positions where the parameter is included in the submodel, and 0s in positions where the parameter is excluded.  This should always be 1 in the positions defining the narrow model, as specified in `inds0`.}

\item{inds0}{Specification of narrow model, in the same format as \code{inds}.  TODO error checking}

\item{gamma0}{Vector of special values taken by the parameters \eqn{gamma} which define the narrow model.

This defaults to 0, as in covariate selection, where excluded coefficients are fixed to 0. 

This should either be a scalar, assumed to be the same for all parameters fixed in the narrow model, or a vector of length equal to the number of parameters from the wide model which are fixed in the narrow model, that is, the number of non-zero entries of inds0.}

\item{focus}{An R function with:

\itemize{
\item first argument named \code{par}, denoting a vector of parameters

\item an optional second argument named \code{X}, typically denoting covariate values.  The required format is documented below. 
}

The function should return the focus quantity of interest.  If \code{X} is supplied and has multiple rows, then \code{focus} should return a vector giving the focus for each row of \code{X}.  Otherwise \code{focus} should return a scalar giving the focus value at \code{par}.

Not required if \code{focus_deriv} is specified.

Alternatively, \code{focus} can be a character string naming a built-in focus function supplied by the \pkg{fic} package.  See \code{\link{focus_fns}}.}

\item{focus_deriv}{Vector of partial derivatives of the focus function with respect to the parameters in the wide model.  This is not usually needed, as it can generally be computed automatically and accurately from the function supplied in \code{focus}, using numerical differentiation.}

\item{X}{Second argument of the focus function, typically giving covariate values defining the focus. This can either be a matrix or a vector.

This is optional.  If not supplied, \code{focus} should return a scalar giving the focus value at parameters \code{par}.

To compute focused model comparison statistics for the same focus function evaluated at multiple covariate values, \code{X} should be a matrix, with number of columns equal to the number of parameters in the wide model, and number of rows equal to the number of alternative covariate values.

For a typical regression model, the first parameter will denote an intercept, so the first value of \code{X} should be 1, and the remaining values should correspond to covariates whose coefficients form parameters of the wide model.  See the examples in the vignette.

If just one covariate value is needed, then \code{X} can be a vector of length equal to the number of parameters in the wide model.}

\item{sub}{List of fitted model objects for each submodel to be assessed.  This is optional, and only required if you want the estimate of the focus function under each submodel to be included in the results.}

\item{fns}{Named list of functions to extract the quantities from the fitted model object that are required for the FIC calculation.  By default this is

\code{list(coef=coef, nobs=nobs, vcov=vcov)}

Suppose the fitted model object is called \code{mod}.  This default list assumes that

\itemize{
\item \code{coef(mod)} returns the vector of parameter estimates,

\item \code{nobs(mod)} returns the number of observations used in the model fit,

\item \code{vcov(mod)} returns the covariance matrix for the parameter estimates.
}

If one or more of these assumptions is not true, then the defaults can be changed.
For example, suppose the functions \code{coef()}, \code{nobs()} and \code{vcov()} are
not understood (or return something different) for your class of model objects, but
the parameters are stored in \code{mod$estimates},
the number of observations is in \code{mod$data$nobs}, and
the covariance matrix is in \code{mod$vcov}, 

then the \code{fns} argument should be set to 

\code{list(
     coef = function(x){x$estimates},
     nobs = function(x){x$data$nobs},
     vcov = function(x){x$cov}
     )}

If less than three components are specified in \code{fns}, then the missing components are assumed to take their default values.}

\item{tidy}{If \code{TRUE} return the results as a data frame.  If \code{FALSE}, return the results as a three-dimensional array, with dimensions indexed by the submodels, result statistics and \code{X} values respectively.}

\item{\dots}{Other arguments to the focus function can be supplied here.  Currently no examples of this.}
}
\value{
A vector containing the following components, describing characteristics of the defined submodel (references in Chapter 6 of Claeskens and Hjort, 2008)

\item{FIC}{The focused information criterion (equation 6.1). }

\item{rmse}{The root mean square error of the estimate of the focus quantity.  Defined on page 150 (equation 6.7).}

\item{rmse.adj}{The root mean square error, with an adjustment to avoid negative squared bias.  Defined on page 157 as the sum of the variance and the squared adjusted bias.}

\item{bias}{The estimated bias of the focus quantity (unadjusted).  Defined on page 157.}

\item{bias.adj}{The estimated bias of the focus quantity (adjusted to avoid negative squared bias).  This is defined as the square root of the quantity "sqb3(S)", page 152, multiplied by the sign of the unadjusted bias, and divided by the square root of the sample size. }

\item{se}{The estimated standard error (root variance) of the focus quantity.  Defined on page 157.}
}
\description{
Focused information criterion for general models.  These methods estimate the bias and variance of estimates of a quantity of interest (the "focus") when parameters are excluded from an "wide" model that is assumed to generate the data.
}
\examples{
TODO

}
\references{
Claeskens, G., & Hjort, N. L. (2008). Model selection and model averaging (Vol. 330). Cambridge: Cambridge University Press.

Claeskens, G., & Hjort, N. L. (2003). The focused information criterion. Journal of the American Statistical Association, 98(464), 900-916.
}
\keyword{models}
