% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fic.R
\name{ggplot_fic}
\alias{ggplot_fic}
\title{Plot focused model comparison statistics: ggplot2 method}
\usage{
ggplot_fic(x, ci = TRUE, ylab = NULL, xlab = NULL, xlim = NULL,
  ylim = NULL)
}
\arguments{
\item{x}{Output from \code{\link{fic}}.}

\item{ci}{Plot interval estimates? (\code{TRUE} or \code{FALSE}).  These are calculated as plus / minus twice the standard error of the submodel focus under the wide model.  These are rough estimates of uncertainty intended to illustrate the bias-variance tradeoff, and exclude any uncertainty associated with the choice between models.}

\item{ylab}{y-axis label.}

\item{xlab}{x-axis label.}

\item{xlim}{x-axis limits (pair of numbers)}

\item{ylim}{y-axis limits}
}
\description{
Plot focused model comparison statistics: ggplot2 method
}
