% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{melanoma}
\alias{melanoma}
\title{Melanoma}
\format{A data frame with 205 rows and the following columns:
\describe{
  \item{ptno}{Patient identification number}
  \item{death}{Survival status: 1 2 3 4 }
  \item{days}{Survival or censoring time}
  \item{depth}{}
  \item{infilt}{}
  \item{mucous}{}
  \item{epith}{}
  \item{ulcer}{}
  \item{thick}{}
  \item{sex}{Sex. Factor with levels \code{"female","male"} and reference level \code{"female"}}
  \item{age}{Age in years}
  \item{qq}{}
  \item{years}{Survival time in years (instead of days)}
  \item{thick_centred}{Centred version of \code{thick}, defined as }
  ...
}}
\source{
The supporting material from Claeskens and Hjort (2008), at \url{https://feb.kuleuven.be/public/u0043181/modelselection/datasets/melanoma_data.txt}.
}
\usage{
melanoma
}
\description{
Melanoma
}
\references{
Claeskens, G., & Hjort, N. L. (2008). Model selection and model averaging (Vol. 330). Cambridge: Cambridge University Press.
}
\keyword{datasets}
