% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{newdata_to_X}
\alias{newdata_to_X}
\title{Convert data frame of covariate values to a design matrix, excluding the intercept}
\usage{
newdata_to_X(newdata, wide)
}
\arguments{
\item{newdata}{Data frame where each row is a vector of covariate values defining an alternative focus quantity.}

\item{wide}{Wide model which includes these covariates.}
}
\value{
"Design" matrix of covariate values defining alternative focuses, with factors expanded to their contrasts.  This is in the form required by the \code{X} argument of \code{\link{fic}}.
}
\description{
Convert data frame of covariate values to a design matrix, excluding the intercept
}
\details{
Numeric values can be supplied for factor levels that are character strings denoting numbers (like \code{"1"} or \code{"2"}).
}
